Pass listen => 1 to Plack::Test::test_psgi
This gets passed on to Test::TCP which asks the kernel for a free
listening socket, avoiding the race conditions inherent in trying to
find a free one manually.

No need to bump the Plack::Test version dependency, since older versions
will just ignore the parameter and keep working in the racy fashion.